requirements.txt
pytest~=8.4.1
selene==2.0.0rc9
selenium~=4.33.0
allure-pytest
allure-python-commons~=2.13.5
python-dotenv~=1.0.1
Faker~=37.4.0
jsonschema
requests